// src/common/constants.d.ts
export declare enum GraphNodeEvent {
    /** 鼠标左键单击节点时触发 */
    onNodeClick = "node:click",
    /** 鼠标双击左键节点时触发 */
    onNodeDoubleClick = "node:dblclick",
    /** 鼠标移入节点时触发 */
    onNodeMouseEnter = "node:mouseenter",
    /** 鼠标在节点内部移到时不断触发，不能通过键盘触发 */
    onNodeMouseMove = "node:mousemove",
    /** 鼠标移出节点后触发 */
    onNodeMouseOut = "node:mouseout",
    /** 鼠标移入节点上方时触发 */
    onNodeMouseOver = "node:mouseover",
    /** 鼠标移出节点时触发 */
    onNodeMouseLeave = "node:mouseleave",
    /** 鼠标按钮在节点上按下（左键或者右键）时触发，不能通过键盘触发 */
    onNodeMouseDown = "node:mousedown",
    /** 节点上按下的鼠标按钮被释放弹起时触发，不能通过键盘触发 */
    onNodeMouseUp = "node:mouseup",
    /** 用户在节点上右击鼠标时触发并打开右键菜单 */
    onNodeContextMenu = "node:contextmenu",
    /** 当节点开始被拖拽的时候触发的事件，此事件作用在被拖曳节点上 */
    onNodeDragStart = "node:dragstart",
    /** 当节点在拖动过程中时触发的事件，此事件作用于被拖拽节点上 */
    onNodeDrag = "node:drag",
    /** 当拖拽完成后触发的事件，此事件作用在被拖曳节点上 */
    onNodeDragEnd = "node:dragend",
    /** 当拖曳节点进入目标元素的时候触发的事件，此事件作用在目标元素上 */
    onNodeDragEnter = "node:dragenter",
    /** 当拖曳节点离开目标元素的时候触发的事件，此事件作用在目标元素上 */
    onNodeDragLeave = "node:dragleave",
    /** 被拖拽的节点在目标元素上同时鼠标放开触发的事件，此事件作用在目标元素上 */
    onNodeDrop = "node:drop"
}
export declare enum GraphComboEvent {
    /** 鼠标左键单击边时触发 */
    onComboClick = "combo:click",
    /** 鼠标双击左键边时触发 */
    onComboDoubleClick = "combo:dblclick",
    /** 鼠标移入边时触发 */
    onComboMouseEnter = "combo:mouseenter",
    /** 鼠标在边上移到时不断触发，不能通过键盘触发 */
    onComboMouseMove = "combo:mousemove",
    /** 鼠标移出边后触发 */
    onComboMouseOut = "combo:mouseout",
    /** 鼠标移入边上方时触发 */
    onComboMouseOver = "combo:mouseover",
    /** 鼠标移出边时触发 */
    onComboMouseLeave = "combo:mouseleave",
    /** 鼠标按钮在边上按下（左键或者右键）时触发，不能通过键盘触发 */
    onComboMouseDown = "combo:mousedown",
    /** 边上按下的鼠标按钮被释放弹起时触发，不能通过键盘触发 */
    onComboMouseUp = "combo:mouseup",
    /** 用户在边上右击鼠标时触发并打开右键菜单 */
    onComboContextMenu = "combo:contextmenu"
}
export declare enum GraphEdgeEvent {
    /** 鼠标左键单击边时触发 */
    onEdgeClick = "edge:click",
    /** 鼠标双击左键边时触发 */
    onEdgeDoubleClick = "edge:dblclick",
    /** 鼠标移入边时触发 */
    onEdgeMouseEnter = "edge:mouseenter",
    /** 鼠标在边上移到时不断触发，不能通过键盘触发 */
    onEdgeMouseMove = "edge:mousemove",
    /** 鼠标移出边后触发 */
    onEdgeMouseOut = "edge:mouseout",
    /** 鼠标移入边上方时触发 */
    onEdgeMouseOver = "edge:mouseover",
    /** 鼠标移出边时触发 */
    onEdgeMouseLeave = "edge:mouseleave",
    /** 鼠标按钮在边上按下（左键或者右键）时触发，不能通过键盘触发 */
    onEdgeMouseDown = "edge:mousedown",
    /** 边上按下的鼠标按钮被释放弹起时触发，不能通过键盘触发 */
    onEdgeMouseUp = "edge:mouseup",
    /** 用户在边上右击鼠标时触发并打开右键菜单 */
    onEdgeContextMenu = "edge:contextmenu"
}